+++ src/components/utils/index.tsx
@@ -6,6 +6,7 @@ import detectSystem from './detectSystem'
 const parseUnit = require('parse-unit')
 
 import type { SdkMetadata, ErrorNames } from '~types/commons'
+import type { TrackedEnvironmentData } from '~types/tracker'
 import type { TranslatedTagParser } from '~types/locales'
 
 export const functionalSwitch = <T extends unknown>(
@@ -203,6 +204,20 @@ export const addDeviceRelatedProperties = (
     isCrossDeviceFlow,
     deviceType: isDesktop ? 'desktop' : 'mobile',
     system,
+  }
+}
+
+export const trackedEnvironmentData = (): TrackedEnvironmentData => {
+  const osInfo = detectSystem('os')
+  const browserInfo = detectSystem('browser')
+
+  return {
+    ...(osInfo && { os: osInfo.name, os_version: osInfo.version }),
+    ...(browserInfo && {
+      browser: browserInfo.name,
+      browser_version: browserInfo.version,
+    }),
+    device: isDesktop ? 'desktop' : 'mobile',
   }
 }